feat(library): Create library, add books
Create a library app. Start it with two initial sample books. Allow the
user to add new books via the "NEW BOOK" button and a form that allows
the user to input the new book's title, author, number of pages, and
whether or not they have read the book. Style the library app with a
library background image and background/button colors that match.
Represent books with cards. Book cards are embedded in a CSS grid with
auto-fit columns so that the "book shelf" will wrap appropriately as
more books are added.